validaciones extra

+++ BibliotecaDiseñoSoftware/src/Dao/DaoLibros.java
@@ -5,6 +5,8 @@
 package Dao;
 
 import Conexion.Conexion_db;
+import Excepciones.AnioSobrepasadoException;
+import Excepciones.CantidadSobrepasadaException;
 import Modelo.Libro;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -19,6 +21,9 @@ import java.util.ArrayList;
 public class DaoLibros {
 
     public void agregarLibro(Libro libro) throws SQLException {
+        if (libro.getAnioPublicacion() > 2023) {
+            throw new AnioSobrepasadoException();
+        }
         try {
             PreparedStatement ps = null;
             Conexion_db conn = new Conexion_db();
@@ -69,6 +74,9 @@ public class DaoLibros {
     }
 
     public void editarLibro(int id, String nombre, String autor, int anioPublicacion, int cantidadCopias, int ID_Genero) throws SQLException {
+        if (anioPublicacion > 2023) {
+            throw new AnioSobrepasadoException();
+        }
         PreparedStatement ps = null;
         Conexion_db obConexion_db = new Conexion_db();
         Connection conn = obConexion_db.getConexion();
@@ -104,33 +112,35 @@ public class DaoLibros {
     }
 
     public void eliminarLibroCantidad(int id, int cantidadIngresada, int cantidadTotal) throws SQLException {
-    PreparedStatement ps = null;
-    Conexion_db obConexion_db = new Conexion_db();
-    Connection conn = obConexion_db.getConexion();
-    try {
-        String sql = "UPDATE libros SET cantidadCopias=? WHERE ID=?";
-        ps = conn.prepareStatement(sql);
-
-        int nuevaCantidad = cantidadTotal - cantidadIngresada;
-
-        if (nuevaCantidad <= 0) {
-            // Si la nueva cantidad es menor o igual a cero, elimina el libro completamente
-            String eliminarSql = "DELETE FROM libros WHERE ID=?";
-            PreparedStatement eliminarPs = conn.prepareStatement(eliminarSql);
-            eliminarPs.setInt(1, id);
-            eliminarPs.executeUpdate();
-        } else {
-            // Actualiza la cantidad de copias
-            ps.setInt(1, nuevaCantidad);
-            ps.setInt(2, id);
-            ps.executeUpdate();
+        if (cantidadIngresada > cantidadTotal) {
+            throw new CantidadSobrepasadaException();
         }
-    } catch (SQLException ex) {
-        ex.printStackTrace();
-        throw new SQLException();
-    }
-}
+        PreparedStatement ps = null;
+        Conexion_db obConexion_db = new Conexion_db();
+        Connection conn = obConexion_db.getConexion();
+        try {
+            String sql = "UPDATE libros SET cantidadCopias=? WHERE ID=?";
+            ps = conn.prepareStatement(sql);
 
+            int nuevaCantidad = cantidadTotal - cantidadIngresada;
+
+            if (nuevaCantidad <= 0) {
+                // Si la nueva cantidad es menor o igual a cero, elimina el libro completamente
+                String eliminarSql = "DELETE FROM libros WHERE ID=?";
+                PreparedStatement eliminarPs = conn.prepareStatement(eliminarSql);
+                eliminarPs.setInt(1, id);
+                eliminarPs.executeUpdate();
+            } else {
+                // Actualiza la cantidad de copias
+                ps.setInt(1, nuevaCantidad);
+                ps.setInt(2, id);
+                ps.executeUpdate();
+            }
+        } catch (SQLException ex) {
+            ex.printStackTrace();
+            throw new SQLException();
+        }
+    }
 
     public ArrayList<Libro> listaLibros() throws SQLException {
